moved post create into a resuable form component
